requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
cached-property==2.0.1
certifi==2025.8.3
charset-normalizer==3.4.3
chevron==0.14.0
click==8.2.1
distro==1.9.0
Flask==3.0.3
frozendict==2.4.6
genanki==0.13.1
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.42.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
